+++ src/main/java/kr/kro/airbob/domain/accommodation/api/AccommodationController.java
@@ -22,7 +22,9 @@ import kr.kro.airbob.cursor.dto.CursorRequest;
 import kr.kro.airbob.domain.accommodation.dto.AccommodationRequest;
 import kr.kro.airbob.domain.accommodation.dto.AccommodationResponse;
 import kr.kro.airbob.domain.accommodation.entity.AccommodationStatus;
-import kr.kro.airbob.domain.accommodation.service.AccommodationService;
+import kr.kro.airbob.domain.accommodation.service.AccommodationCommandService;
+import kr.kro.airbob.domain.accommodation.service.AccommodationImageService;
+import kr.kro.airbob.domain.accommodation.service.AccommodationQueryService;
 import kr.kro.airbob.domain.auth.annotation.CurrentMemberId;
 import kr.kro.airbob.domain.image.dto.ImageResponse;
 import lombok.RequiredArgsConstructor;
@@ -34,12 +36,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequestMapping("/api")
 public class AccommodationController {
 
-    private final AccommodationService accommodationService;
+    private final AccommodationCommandService accommodationCommandService;
+    private final AccommodationImageService accommodationImageService;
+    private final AccommodationQueryService accommodationQueryService;
 
     @PostMapping("/v1/accommodations")
     public ResponseEntity<ApiResponse<AccommodationResponse.Create>> registerAccommodation(
         @CurrentMemberId Long memberId) {
-        AccommodationResponse.Create response = accommodationService.createAccommodation(memberId);
+        AccommodationResponse.Create response = accommodationCommandService.createAccommodation(memberId);
         return ResponseEntity.status(HttpStatus.CREATED)
             .body(ApiResponse.success(response));
     }
@@ -48,14 +52,14 @@ public class AccommodationController {
     public ResponseEntity<ApiResponse<Void>> updateAccommodation(@PathVariable Long accommodationId,
         @RequestBody @Valid AccommodationRequest.Update request,
         @CurrentMemberId Long memberId) {
-        accommodationService.updateAccommodation(accommodationId, request, memberId);
+        accommodationCommandService.updateAccommodation(accommodationId, request, memberId);
         return ResponseEntity.ok(ApiResponse.success());
     }
 
     @PatchMapping("/v1/accommodations/{accommodationId}/publish")
     public ResponseEntity<ApiResponse<Void>> publishAccommodation(@PathVariable Long accommodationId,
         @CurrentMemberId Long memberId) {
-        accommodationService.publishAccommodation(accommodationId, memberId);
+        accommodationCommandService.publishAccommodation(accommodationId, memberId);
 
         return ResponseEntity.ok(ApiResponse.success());
     }
@@ -63,7 +67,7 @@ public class AccommodationController {
     @PatchMapping("/v1/accommodations/{accommodationId}/unpublish")
     public ResponseEntity<ApiResponse<Void>> unpublishAccommodation(@PathVariable Long accommodationId,
         @CurrentMemberId Long memberId) {
-        accommodationService.unpublishAccommodation(accommodationId, memberId);
+        accommodationCommandService.unpublishAccommodation(accommodationId, memberId);
 
         return ResponseEntity.ok(ApiResponse.success());
     }
@@ -71,7 +75,7 @@ public class AccommodationController {
     @DeleteMapping("/v1/accommodations/{accommodationId}")
     public ResponseEntity<ApiResponse<Void>> deleteAccommodation(@PathVariable Long accommodationId,
         @CurrentMemberId Long memberId) {
-        accommodationService.deleteAccommodation(accommodationId, memberId);
+        accommodationCommandService.deleteAccommodation(accommodationId, memberId);
         return ResponseEntity.ok(ApiResponse.success());
     }
 
@@ -80,7 +84,7 @@ public class AccommodationController {
         @PathVariable Long accommodationId,
         @RequestParam("images") List<MultipartFile> images,
         @CurrentMemberId Long memberId) {
-        ImageResponse.ImageUploadResult response = accommodationService.uploadImages(accommodationId, images,
+        ImageResponse.ImageUploadResult response = accommodationImageService.uploadImages(accommodationId, images,
             memberId);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.success(response));
@@ -91,7 +95,7 @@ public class AccommodationController {
         @PathVariable Long accommodationId,
         @PathVariable Long imageId,
         @CurrentMemberId Long memberId) {
-        accommodationService.deleteImage(accommodationId, imageId, memberId);
+        accommodationImageService.deleteImage(accommodationId, imageId, memberId);
 
         return ResponseEntity.ok(ApiResponse.success());
     }
@@ -100,7 +104,7 @@ public class AccommodationController {
     public ResponseEntity<ApiResponse<AccommodationResponse.DetailInfo>> getAccommodation(
         @PathVariable Long accommodationId,
         @CurrentMemberId(required = false) Long viewerId) {
-        AccommodationResponse.DetailInfo response = accommodationService.findAccommodation(accommodationId,
+        AccommodationResponse.DetailInfo response = accommodationQueryService.findAccommodation(accommodationId,
             viewerId);
         return ResponseEntity.ok(ApiResponse.success(response));
     }
@@ -110,7 +114,7 @@ public class AccommodationController {
         @CursorParam CursorRequest.CursorPageRequest request,
         @RequestParam(required = false) AccommodationStatus status,
         @CurrentMemberId Long memberId) {
-        AccommodationResponse.HostAccommodationInfos response = accommodationService.findMyAccommodations(
+        AccommodationResponse.HostAccommodationInfos response = accommodationQueryService.findMyAccommodations(
             memberId, request, status);
         return ResponseEntity.ok(ApiResponse.success(response));
     }
@@ -119,7 +123,8 @@ public class AccommodationController {
     public ResponseEntity<ApiResponse<AccommodationResponse.HostDetail>> getHostAccommodationDetail(
         @PathVariable Long accommodationId,
         @CurrentMemberId Long hostId) {
-        AccommodationResponse.HostDetail response = accommodationService.findHostAccommodationDetail(accommodationId, hostId);
+        AccommodationResponse.HostDetail response = accommodationQueryService.findHostAccommodationDetail(
+            accommodationId, hostId);
 
         return ResponseEntity.ok(ApiResponse.success(response));
     }
