--- src/main/java/kr/kro/airbob/domain/accommodation/service/AccommodationService.java
@@ -1,629 +1,0 @@
-package kr.kro.airbob.domain.accommodation.service;
-
-import static kr.kro.airbob.search.event.AccommodationIndexingEvents.*;
-
-import java.io.IOException;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.AbstractMap;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Slice;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import kr.kro.airbob.common.exception.InvalidInputException;
-import kr.kro.airbob.common.history.ChangeType;
-import kr.kro.airbob.common.history.HistoryConstants;
-import kr.kro.airbob.cursor.dto.CursorRequest;
-import kr.kro.airbob.cursor.dto.CursorResponse;
-import kr.kro.airbob.cursor.util.CursorPageInfoCreator;
-import kr.kro.airbob.domain.commoncode.common.CommonCodeGroups;
-import kr.kro.airbob.domain.commoncode.service.CommonCodeService;
-import kr.kro.airbob.domain.accommodation.dto.AccommodationRequest;
-import kr.kro.airbob.domain.accommodation.dto.AccommodationResponse;
-import kr.kro.airbob.domain.accommodation.dto.AddressRequest;
-import kr.kro.airbob.domain.accommodation.dto.AmenityRequest;
-import kr.kro.airbob.domain.accommodation.dto.AmenityResponse;
-import kr.kro.airbob.domain.accommodation.dto.PolicyRequest;
-import kr.kro.airbob.domain.accommodation.entity.Accommodation;
-import kr.kro.airbob.domain.accommodation.entity.AccommodationAmenity;
-import kr.kro.airbob.domain.accommodation.entity.AccommodationHistory;
-import kr.kro.airbob.domain.accommodation.entity.AccommodationStatus;
-import kr.kro.airbob.domain.accommodation.entity.Address;
-import kr.kro.airbob.domain.accommodation.entity.OccupancyPolicy;
-import kr.kro.airbob.domain.accommodation.exception.AccommodationNotFoundException;
-import kr.kro.airbob.domain.accommodation.exception.AccommodationStateException;
-import kr.kro.airbob.domain.accommodation.exception.InvalidAccommodationAmenityException;
-import kr.kro.airbob.domain.accommodation.exception.InvalidAccommodationTypeException;
-import kr.kro.airbob.domain.accommodation.exception.PublishingFieldRequiredException;
-import kr.kro.airbob.domain.accommodation.repository.AccommodationAmenityRepository;
-import kr.kro.airbob.domain.accommodation.repository.AccommodationHistoryRepository;
-import kr.kro.airbob.domain.accommodation.repository.AccommodationImageRepository;
-import kr.kro.airbob.domain.accommodation.repository.AccommodationRepository;
-import kr.kro.airbob.domain.accommodation.repository.AddressRepository;
-import kr.kro.airbob.domain.accommodation.repository.OccupancyPolicyRepository;
-import kr.kro.airbob.domain.accommodation.repository.projection.AccommodationDetailProjection;
-import kr.kro.airbob.domain.image.dto.ImageResponse;
-import kr.kro.airbob.domain.image.entity.AccommodationImage;
-import kr.kro.airbob.domain.image.exception.EmptyImageFileException;
-import kr.kro.airbob.domain.image.exception.ImageFileSizeExceededException;
-import kr.kro.airbob.domain.image.exception.ImageNotFoundException;
-import kr.kro.airbob.domain.image.exception.ImageUploadException;
-import kr.kro.airbob.domain.image.exception.InsufficientImageCountException;
-import kr.kro.airbob.domain.image.exception.InvalidImageFormatException;
-import kr.kro.airbob.domain.image.service.S3ImageUploader;
-import kr.kro.airbob.domain.member.entity.Member;
-import kr.kro.airbob.domain.member.entity.MemberStatus;
-import kr.kro.airbob.domain.member.exception.MemberNotFoundException;
-import kr.kro.airbob.domain.member.repository.MemberRepository;
-import kr.kro.airbob.domain.reservation.dto.ReservationDateRange;
-import kr.kro.airbob.domain.reservation.policy.BookingWindow;
-import kr.kro.airbob.domain.reservation.repository.ReservationRepository;
-import kr.kro.airbob.domain.review.dto.ReviewResponse;
-import kr.kro.airbob.domain.review.entity.AccommodationReviewSummary;
-import kr.kro.airbob.domain.review.repository.AccommodationReviewSummaryRepository;
-import kr.kro.airbob.domain.wishlist.repository.WishlistAccommodationRepository;
-import kr.kro.airbob.geo.GeocodingService;
-import kr.kro.airbob.geo.dto.GeocodeResult;
-import kr.kro.airbob.outbox.EventType;
-import kr.kro.airbob.outbox.OutboxEventPublisher;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class AccommodationService {
-
-    public static final int MAX_IMAGE_SIZE = 10 * 1024 * 1024;
-    public static final String IMAGE_JPEG = "image/jpeg";
-    public static final String IMAGE_PNG = "image/png";
-    private final AccommodationReviewSummaryRepository reviewSummaryRepository;
-    private final WishlistAccommodationRepository wishlistAccommodationRepository;
-    private final AccommodationAmenityRepository accommodationAmenityRepository;
-    private final AccommodationImageRepository accommodationImageRepository;
-    private final AccommodationHistoryRepository accommodationHistoryRepository;
-    private final OccupancyPolicyRepository occupancyPolicyRepository;
-    private final AccommodationRepository accommodationRepository;
-    private final ReservationRepository reservationRepository;
-    private final CommonCodeService commonCodeService;
-    private final AddressRepository addressRepository;
-    private final MemberRepository memberRepository;
-
-    private final CursorPageInfoCreator cursorPageInfoCreator;
-    private final OutboxEventPublisher outboxEventPublisher;
-    private final GeocodingService geocodingService;
-    private final S3ImageUploader s3ImageUploader;
-
-    @Transactional
-    public AccommodationResponse.Create createAccommodation(Long memberId) {
-
-        Member member = memberRepository.findByIdAndStatus(memberId, MemberStatus.ACTIVE)
-                .orElseThrow(MemberNotFoundException::new);
-
-        Accommodation accommodation = Accommodation.createAccommodation(member);
-        Accommodation savedAccommodation = accommodationRepository.save(accommodation);
-
-        recordHistory(savedAccommodation, ChangeType.CREATE, "숙소 생성");
-
-        return new AccommodationResponse.Create(savedAccommodation.getId());
-    }
-
-    @Transactional
-    public void updateAccommodation(Long accommodationId, AccommodationRequest.Update request, Long memberId) {
-
-        Accommodation accommodation = findByIdAndMemberIdAndStatusNot(accommodationId, memberId);
-
-        validateAccommodationType(request.type());
-        Map<String, Integer> amenityCountMap = getAmenityCountMap(request.amenityInfos());
-        accommodation.updateAccommodation(request);
-        updateAddress(accommodation, request.addressInfo());
-        updateOccupancyPolicy(accommodation, request.occupancyPolicyInfo());
-        updateAmenities(accommodation, amenityCountMap);
-
-        recordHistory(accommodation, ChangeType.UPDATE, "숙소 정보 수정");
-
-        if (accommodation.getStatus() == AccommodationStatus.PUBLISHED) { // 게시 상태인 경우 ES 색인 이벤트 발행
-            outboxEventPublisher.save(
-                EventType.ACCOMMODATION_UPDATED,
-                new AccommodationUpdatedEvent(accommodation.getAccommodationUid().toString())
-            );
-        }
-    }
-
-    @Transactional
-    public void deleteAccommodation(Long accommodationId, Long memberId) {
-        Accommodation accommodation = findByIdAndMemberIdExceptDeleted(accommodationId, memberId);
-
-        accommodation.delete();
-
-        recordHistory(accommodation, ChangeType.DELETE, "숙소 삭제");
-
-        outboxEventPublisher.save(
-            EventType.ACCOMMODATION_DELETED,
-            new AccommodationDeletedEvent(accommodation.getAccommodationUid().toString())
-        );
-    }
-
-    @Transactional(readOnly = true)
-    public AccommodationResponse.DetailInfo findAccommodation(Long accommodationId, Long viewerId) {
-
-        // 숙소 본문 + 리뷰 요약
-        AccommodationDetailProjection detailProjection = accommodationRepository
-            .findWithDetailsByAccommodationIdAndStatus(accommodationId, AccommodationStatus.PUBLISHED)
-            .orElseThrow(AccommodationNotFoundException::new);
-        Accommodation accommodation = detailProjection.accommodation();
-
-        // 편의 시설
-        List<AmenityResponse.AmenityInfo> amenityInfos = getAmenities(accommodationId);
-
-        // 이미지
-        List<ImageResponse.ImageInfo> imageInfos = getImageUrls(accommodationId);
-
-        // 예약 불가 날짜 (3개월)
-        BookingWindow bookingWindow = BookingWindow.current();
-        LocalDate bookingWindowStart = bookingWindow.startInclusive();
-        LocalDate bookingWindowEndExclusive = bookingWindow.endExclusive();
-        List<AccommodationResponse.UnavailableDateRange> unavailableRanges = getUnavailableRanges(
-            accommodationId, bookingWindowStart, bookingWindowEndExclusive);
-
-        // 위시리스트 포함 여부 - 로그인 사용자만
-        Boolean isInWishlist = checkWishlistStatus(accommodationId, viewerId);
-        ReviewResponse.ReviewSummary reviewSummary = new ReviewResponse.ReviewSummary(
-            Objects.requireNonNullElse(detailProjection.totalReviewCount(), 0),
-            Objects.requireNonNullElse(detailProjection.averageRating(), BigDecimal.ZERO)
-        );
-
-        return AccommodationResponse.DetailInfo.from(
-            accommodation, bookingWindowStart, bookingWindowEndExclusive, unavailableRanges, isInWishlist,
-            amenityInfos, imageInfos, reviewSummary);
-    }
-
-    @Transactional(readOnly = true)
-    public AccommodationResponse.HostAccommodationInfos findMyAccommodations(Long hostId, CursorRequest.CursorPageRequest cursorRequest, AccommodationStatus status) {
-        Slice<Accommodation> accommodationSlice = accommodationRepository.findMyAccommodationsByHostIdWithCursor(
-            hostId,
-            cursorRequest.lastId(),
-            cursorRequest.lastCreatedAt(),
-            status,
-            PageRequest.of(0, cursorRequest.size())
-        );
-
-        List<Accommodation> accommodations = accommodationSlice.getContent();
-        if (accommodations.isEmpty()) {
-            CursorResponse.PageInfo pageInfo = cursorPageInfoCreator.createPageInfo(
-                Collections.emptyList(), false, acc -> 0L, acc -> null
-            );
-            return AccommodationResponse.HostAccommodationInfos.builder()
-                .accommodations(Collections.emptyList())
-                .pageInfo(pageInfo)
-                .build();
-        }
-
-        List<AccommodationResponse.HostAccommodationInfo> accommodationInfos =
-            accommodations.stream()
-                .map(AccommodationResponse.HostAccommodationInfo::from)
-                .toList();
-
-        CursorResponse.PageInfo pageInfo = cursorPageInfoCreator.createPageInfo(
-            accommodations,
-            accommodationSlice.hasNext(),
-            Accommodation::getId,
-            Accommodation::getCreatedAt
-        );
-
-        return AccommodationResponse.HostAccommodationInfos.from(accommodationInfos, pageInfo);
-    }
-
-    @Transactional
-    public ImageResponse.ImageUploadResult uploadImages(Long accommodationId, List<MultipartFile> images,
-        Long memberId) {
-
-        // todo: 숙소 + 호스트 조회 -> 호스트인지 여부 검증으로 변경 필요
-        Accommodation accommodation = findByIdAndMemberIdExceptDeleted(accommodationId, memberId);
-
-        List<ImageResponse.ImageInfo> imageInfos = new ArrayList<>();
-        List<AccommodationImage> savedImages = new ArrayList<>();
-
-        for (MultipartFile image : images) {
-            validateImageFile(image);
-
-            String imageUrl;
-            try {
-                String dirName = "accommodationInfos/" + accommodationId;
-                imageUrl = s3ImageUploader.upload(image, dirName);
-            } catch (IOException e) {
-                log.error("이미지 업로드 실패: accommodationId={}, fileName={}", accommodation.getId(),
-                    image.getOriginalFilename(), e);
-                throw new ImageUploadException(image.getOriginalFilename());
-            }
-
-            AccommodationImage accommodationImage = AccommodationImage.builder()
-                .accommodation(accommodation)
-                .imageUrl(imageUrl)
-                .build();
-            savedImages.add(accommodationImage);
-        }
-
-        List<AccommodationImage> actuallySavedImages = accommodationImageRepository.saveAll(savedImages);
-
-        for (AccommodationImage savedImage : actuallySavedImages) {
-            imageInfos.add(ImageResponse.ImageInfo.builder()
-                .id(savedImage.getId())
-                .imageUrl(savedImage.getImageUrl())
-                .build());
-        }
-
-        findAndUpdateThumbnail(accommodation);
-
-        return ImageResponse.ImageUploadResult.from(imageInfos);
-    }
-
-    //todo: query 2번
-    @Transactional
-    public void deleteImage(Long accommodationId, Long imageId, Long memberId) {
-        Accommodation accommodation = findByIdAndMemberIdExceptDeleted(accommodationId, memberId);
-
-        // 숙소 이미지 + 숙소 함께 조회
-        AccommodationImage image = accommodationImageRepository.findById(imageId)
-            .orElseThrow(ImageNotFoundException::new);
-
-        if (!image.getAccommodation().getId().equals(accommodationId)) {
-            throw new InvalidInputException();
-        }
-
-        s3ImageUploader.delete(image.getImageUrl());
-
-        accommodationImageRepository.delete(image);
-
-        if (image.getImageUrl().equals(accommodation.getThumbnailUrl())) {
-            // 삭제된 이미지가 현재 썸네일이었을 경우 새로운 썸네일 설정
-            findAndUpdateThumbnail(accommodation);
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public AccommodationResponse.HostDetail findHostAccommodationDetail(Long accommodationId, Long hostId) {
-        Accommodation accommodation = accommodationRepository.findWithDetailsByIdAndHostId(accommodationId, hostId)
-            .orElseThrow(AccommodationNotFoundException::new);
-
-        List<AmenityResponse.AmenityInfo> amenityInfos = getAmenities(accommodationId);
-        List<ImageResponse.ImageInfo> imageInfos = getImageUrls(accommodationId);
-        ReviewResponse.ReviewSummary reviewSummary = getReviewSummary(accommodationId);
-
-        return AccommodationResponse.HostDetail.from(
-            accommodation,
-            amenityInfos,
-            imageInfos,
-            reviewSummary
-        );
-    }
-
-    @Transactional
-    public void publishAccommodation(Long accommodationId, Long memberId) {
-        Accommodation accommodation = findWithDetailsExceptHostAndDeletedById(accommodationId, memberId);
-
-        validateAccommodationForPublishing(accommodation);
-
-        accommodation.publish();
-
-        recordHistory(accommodation, ChangeType.STATUS_CHANGE, "숙소 게시");
-
-        outboxEventPublisher.save(
-            EventType.ACCOMMODATION_UPDATED,
-            new AccommodationUpdatedEvent(accommodation.getAccommodationUid().toString())
-        );
-    }
-
-    @Transactional
-    public void unpublishAccommodation(Long accommodationId, Long memberId) {
-        Accommodation accommodation = findByIdAndMemberIdExceptDeleted(accommodationId, memberId);
-
-        if (accommodation.getStatus() != AccommodationStatus.PUBLISHED) {
-            throw new AccommodationStateException();
-        }
-
-        accommodation.unpublish();
-
-        recordHistory(accommodation, ChangeType.STATUS_CHANGE, "숙소 게시 중단");
-
-        outboxEventPublisher.save(
-            EventType.ACCOMMODATION_UPDATED,
-            new AccommodationUpdatedEvent(accommodation.getAccommodationUid().toString())
-        );
-    }
-
-    // 이력 기록(SCD2): 직전 현재 행을 닫고 현재 상태의 전체 스냅샷을 새로 연다
-    private void recordHistory(Accommodation accommodation, ChangeType changeType, String reason) {
-        accommodationHistoryRepository.findByAccommodationIdAndValidTo(accommodation.getId(), HistoryConstants.FOREVER)
-            .ifPresent(current -> current.close(LocalDateTime.now()));
-        accommodationHistoryRepository.save(AccommodationHistory.of(accommodation, changeType, reason));
-    }
-
-    private void validateAccommodationForPublishing(Accommodation accommodation) {
-        if (accommodation.getName() == null || accommodation.getName().isBlank()) {
-            throw new PublishingFieldRequiredException("name");
-        }
-        if (accommodation.getDescription() == null || accommodation.getDescription().isBlank()) {
-            throw new PublishingFieldRequiredException("description");
-        }
-
-        Address address = accommodation.getAddress();
-        if (address == null
-            || address.getCountry() == null
-            || address.getCity() == null
-            || address.getStreet() == null
-            || address.getPostalCode() == null
-            || address.getLatitude() == null
-            || address.getLongitude() == null) {
-
-            throw new PublishingFieldRequiredException("addressInfo", "주소 정보(세부 항목 포함)가 누락되었습니다.");
-        }
-
-        if (accommodation.getBasePrice() == null || accommodation.getBasePrice() < 1) { // 가격 검증 (1 이상)
-            throw new PublishingFieldRequiredException("basePrice", "기본 가격은 1원 이상이어야 합니다.");
-        }
-
-        if (accommodation.getType() == null) {
-            throw new PublishingFieldRequiredException("type");
-        }
-        OccupancyPolicy policy = accommodation.getOccupancyPolicy();
-        if (policy == null || policy.getMaxOccupancy() == null
-            || policy.getInfantOccupancy() == null
-            || policy.getPetOccupancy() == null) {
-
-            throw new PublishingFieldRequiredException("occupancyPolicyInfo", "수용 인원 정책(세부 항목 포함)이 누락되었습니다.");
-        }
-
-        if (accommodation.getCheckInTime() == null || accommodation.getCheckOutTime() == null) {
-            throw new PublishingFieldRequiredException("checkInTime/checkOutTime", "체크인/체크아웃 시간은 필수입니다.");
-        }
-
-        long imageCount = accommodationImageRepository.countByAccommodationId(accommodation.getId());
-        int minImageCount = 1; // 최소 이미지 개수
-        if (imageCount < minImageCount) {
-            throw new InsufficientImageCountException("이미지는 최소 " + minImageCount + "개 이상 등록해야 게시할 수 있습니다. 현재: " + imageCount + "개");
-        }
-    }
-
-    private void validateImageFile(MultipartFile file) {
-        if (file.isEmpty()) {
-            throw new EmptyImageFileException();
-        }
-
-        if (file.getSize() > MAX_IMAGE_SIZE) {
-            throw new ImageFileSizeExceededException();
-        }
-
-        String contentType = file.getContentType();
-        if (contentType == null || (!contentType.equals(IMAGE_JPEG) && !contentType.equals(IMAGE_PNG))) {
-            throw new InvalidImageFormatException();
-        }
-    }
-
-    private void findAndUpdateThumbnail(Accommodation accommodation) {
-
-        List<AccommodationImage> remainingImages = accommodationImageRepository.findByAccommodationIdOrderByIdAsc(accommodation.getId());
-        String newThumbnailUrl = remainingImages.isEmpty() ? null : remainingImages.getFirst().getImageUrl();
-
-        if (!Objects.equals(accommodation.getThumbnailUrl(), newThumbnailUrl)) {
-            accommodation.updateThumbnailUrl(newThumbnailUrl);
-        }
-    }
-
-    private List<AmenityResponse.AmenityInfo> getAmenities(Long accommodationId) {
-        return accommodationAmenityRepository.findAllByAccommodationId(accommodationId)
-            .stream()
-            .map(AmenityResponse.AmenityInfo::from)
-            .toList();
-    }
-
-    private List<ImageResponse.ImageInfo> getImageUrls(Long accommodationId) {
-        return accommodationImageRepository.findByAccommodationIdOrderByIdAsc(accommodationId)
-            .stream()
-            .map(ImageResponse.ImageInfo::from)
-            .toList();
-    }
-
-    private ReviewResponse.ReviewSummary getReviewSummary(Long accommodationId) {
-        Optional<AccommodationReviewSummary> summaryOpt = reviewSummaryRepository.findByAccommodationId(
-            accommodationId);
-        return ReviewResponse.ReviewSummary.of(summaryOpt.orElse(null));
-    }
-
-    private List<AccommodationResponse.UnavailableDateRange> getUnavailableRanges(
-        Long accommodationId,
-        LocalDate windowStart,
-        LocalDate windowEndExclusive
-    ) {
-        List<ReservationDateRange> reservationRanges = reservationRepository
-            .findConfirmedReservationRangesByAccommodationId(
-                accommodationId,
-                windowStart.atStartOfDay(),
-                windowEndExclusive.atStartOfDay());
-
-        List<AccommodationResponse.UnavailableDateRange> clippedRanges = reservationRanges.stream()
-            .map(range -> clipUnavailableRange(range, windowStart, windowEndExclusive))
-            .filter(range -> range.startDate().isBefore(range.endDateExclusive()))
-            .sorted(Comparator
-                .comparing(AccommodationResponse.UnavailableDateRange::startDate)
-                .thenComparing(AccommodationResponse.UnavailableDateRange::endDateExclusive))
-            .toList();
-
-        return mergeUnavailableRanges(clippedRanges);
-    }
-
-    private AccommodationResponse.UnavailableDateRange clipUnavailableRange(
-        ReservationDateRange range,
-        LocalDate windowStart,
-        LocalDate windowEndExclusive
-    ) {
-        LocalDate startDate = range.checkIn().toLocalDate();
-        LocalDate endDateExclusive = range.checkOut().toLocalDate();
-
-        if (startDate.isBefore(windowStart)) {
-            startDate = windowStart;
-        }
-        if (endDateExclusive.isAfter(windowEndExclusive)) {
-            endDateExclusive = windowEndExclusive;
-        }
-
-        return new AccommodationResponse.UnavailableDateRange(startDate, endDateExclusive);
-    }
-
-    private List<AccommodationResponse.UnavailableDateRange> mergeUnavailableRanges(
-        List<AccommodationResponse.UnavailableDateRange> ranges
-    ) {
-        if (ranges.isEmpty()) {
-            return List.of();
-        }
-
-        List<AccommodationResponse.UnavailableDateRange> mergedRanges = new ArrayList<>();
-        AccommodationResponse.UnavailableDateRange current = ranges.getFirst();
-
-        for (int index = 1; index < ranges.size(); index++) {
-            AccommodationResponse.UnavailableDateRange next = ranges.get(index);
-
-            if (!next.startDate().isAfter(current.endDateExclusive())) {
-                LocalDate mergedEnd = next.endDateExclusive().isAfter(current.endDateExclusive())
-                    ? next.endDateExclusive()
-                    : current.endDateExclusive();
-                current = new AccommodationResponse.UnavailableDateRange(current.startDate(), mergedEnd);
-                continue;
-            }
-
-            mergedRanges.add(current);
-            current = next;
-        }
-
-        mergedRanges.add(current);
-        return List.copyOf(mergedRanges);
-    }
-
-    private Boolean checkWishlistStatus(Long accommodationId, Long viewerId) {
-        if (viewerId == null) {
-            return false; // 비로그인은 false
-        }
-
-        return wishlistAccommodationRepository.existsByWishlist_Member_IdAndAccommodation_Id(viewerId,
-            accommodationId);
-    }
-
-
-    private void saveValidAmenities(Map<String, Integer> amenityCountMap, Accommodation savedAccommodation) {
-        if(amenityCountMap.isEmpty()) return;
-
-        // 공통 코드(AMENITY_TYPE)로 이미 검증된 코드만 남으므로, 코드 문자열을 직접 저장(amenity 테이블 조회 불필요)
-        List<AccommodationAmenity> accommodationAmenityList = amenityCountMap.entrySet().stream()
-            .map(entry -> AccommodationAmenity.createAccommodationAmenity(
-                savedAccommodation, entry.getKey(), entry.getValue()))
-            .toList();
-
-        accommodationAmenityRepository.saveAll(accommodationAmenityList);
-    }
-
-    // 편의시설 코드 정합성은 공통 코드(AMENITY_TYPE) 캐시로 검증 — enum 대신 애플리케이션 레벨 통제
-    private Map<String, Integer> getAmenityCountMap(List<AmenityRequest.AmenityInfo> request) {
-        if (request == null) {
-            return null;
-        }
-
-        Map<String, Integer> amenityCountMap = request.stream()
-            .filter(info -> info.count() > 0)
-            .map(info -> new AbstractMap.SimpleEntry<>(
-                info.name() == null ? null : info.name().toUpperCase(Locale.ROOT), info.count()))
-            .collect(Collectors.toMap(
-                Map.Entry::getKey,
-                Map.Entry::getValue,
-                Integer::sum
-            ));
-
-        boolean hasInvalidCode = amenityCountMap.keySet().stream()
-            .anyMatch(code -> !commonCodeService.isValidCode(CommonCodeGroups.AMENITY_TYPE, code));
-        if (hasInvalidCode) {
-            throw new InvalidAccommodationAmenityException();
-        }
-
-        return amenityCountMap;
-    }
-
-    // 숙소 유형 코드 정합성 검증(공통 코드 ACCOMMODATION_TYPE). 유효하지 않으면 거부.
-    private void validateAccommodationType(String type) {
-        if (type == null) {
-            return;
-        }
-        if (!commonCodeService.isValidCode(
-            CommonCodeGroups.ACCOMMODATION_TYPE, type.toUpperCase(Locale.ROOT))) {
-            throw new InvalidAccommodationTypeException();
-        }
-    }
-
-    private void updateOccupancyPolicy(Accommodation accommodation, PolicyRequest.OccupancyPolicyInfo occupancyPolicyInfo) {
-        if (occupancyPolicyInfo == null) {
-            return;
-        }
-
-        OccupancyPolicy currentPolicy = accommodation.getOccupancyPolicy();
-        if (currentPolicy == null) {
-            OccupancyPolicy newPolicy = OccupancyPolicy.createOccupancyPolicy(occupancyPolicyInfo);
-            OccupancyPolicy savedPolicy = occupancyPolicyRepository.save(newPolicy);
-            accommodation.updateOccupancyPolicy(savedPolicy);
-        }
-    }
-
-    private void updateAddress(Accommodation accommodation, AddressRequest.AddressInfo addressInfo) {
-        if (addressInfo == null) {
-            return;
-        }
-
-        Address currentAddress = accommodation.getAddress();
-
-        if (currentAddress == null || currentAddress.isChanged(addressInfo)) {
-            String addressStr = Address.buildAddressStringForGeocoding(addressInfo);
-            GeocodeResult geocodeResult = geocodingService.getCoordinates(addressStr);
-            Address newAddress = Address.createAddress(addressInfo, geocodeResult);
-            Address savedAddress = addressRepository.save(newAddress);
-            accommodation.updateAddress(savedAddress);
-        }
-    }
-
-    private void updateAmenities(Accommodation accommodation, Map<String, Integer> amenityCountMap) {
-        if (amenityCountMap == null) {
-            return;
-        }
-
-        accommodationAmenityRepository.deleteByAccommodationIdInBulk(accommodation.getId());
-
-        if (!amenityCountMap.isEmpty()) {
-            saveValidAmenities(amenityCountMap, accommodation);
-        }
-    }
-
-    private Accommodation findWithDetailsExceptHostAndDeletedById(Long accommodationId, Long memberId) {
-        return accommodationRepository.findWithDetailsExceptHostAndDeletedById(accommodationId, memberId).orElseThrow(AccommodationNotFoundException::new);
-    }
-
-    private Accommodation findByIdAndMemberIdAndStatusNot(Long accommodationId, Long memberId) {
-        return accommodationRepository.findByIdAndMemberIdAndStatusNot(accommodationId, memberId, AccommodationStatus.DELETED).orElseThrow(AccommodationNotFoundException::new);
-    }
-
-    private Accommodation findByIdAndStatus(Long accommodationId, AccommodationStatus status) {
-        return accommodationRepository.findByIdAndStatus(accommodationId, status).orElseThrow(AccommodationNotFoundException::new);
-    }
-
-    private Accommodation findByIdAndMemberIdExceptDeleted(Long accommodationId, Long memberId) {
-        return accommodationRepository.findByIdAndMemberIdAndStatusNot(accommodationId, memberId, AccommodationStatus.DELETED).orElseThrow(AccommodationNotFoundException::new);
-    }
-}
